Initial import
This is a tool to create a vdisc csv manifest file from a given set
of Swift containers. It uses the python-swiftclient library to auth,
query the cluster, and generate the appropriate listings.

Usage is ./create_manifest.py c1 [c2 [c3 [...]]]

Redirect the output to a csv file to use it with vdisk.

The current vdisc swift url is:
  swift://<endpoint>/<container_owner>/<container>/path/to/some.obj

Signed-off-by: John Dickinson <me@not.mn>